✨ feat(ar-generator-service): Rename the database connection module for clarity
